+++ actions/ReportsList.php
@@ -4,6 +4,9 @@ namespace Modules\MoreReporting\Actions;
 
 use CController, CControllerResponseData;
 
+use Modules\MoreReporting\Includes\ReportStorage;
+use Modules\MoreReporting\Includes\ReportTypeRegistry;
+
 class ReportsList extends CController {
 
 	public function init(): void {
@@ -21,6 +24,8 @@ class ReportsList extends CController {
 	protected function doAction(): void {
 		$data = [
 			'title' => _('MoreReporting'),
+			'reports' => ReportStorage::getAll(),
+			'report_type_labels' => ReportTypeRegistry::labels()
 		];
 
 		$response = new CControllerResponseData($data);